+++ android/src/main/java/map/naver/plugin/net/lbstech/naver_map_plugin/NaverMarkerController.java
@@ -74,7 +74,8 @@ class NaverMarkerController {
         for (Object json : jsonArray) {
             HashMap<String, Object> data = (HashMap<String, Object>) json;
             String id = (String) data.get("markerId");
-            idToController.get(id).interpret(data);
+            if (idToController.containsKey(id) && idToController.get(id) != null)
+                idToController.get(id).interpret(data);
         }
     }
 
@@ -98,7 +99,7 @@ class NaverMarkerController {
             if (position != null) marker.setPosition(Convert.toLatLng(position));
 
             final Object alpha = json.get("alpha");
-            if (alpha != null) marker.setAlpha((float) alpha);
+            if (alpha != null) marker.setAlpha(Convert.toFloat(alpha));
 
             final Object flat = json.get("flat");
             if (flat != null) marker.setFlat((boolean) flat);
@@ -107,7 +108,7 @@ class NaverMarkerController {
             if (captionText != null) marker.setCaptionText((String) captionText);
 
             final Object captionTextSize = json.get("captionTextSize");
-            if (captionTextSize != null) marker.setCaptionTextSize((float) captionTextSize);
+            if (captionTextSize != null) marker.setCaptionTextSize(Convert.toFloat(captionTextSize));
 
             final Object captionColor = json.get("captionColor");
             if (captionColor != null) marker.setCaptionColor(Convert.toColorInt(captionColor));
@@ -159,7 +160,7 @@ class NaverMarkerController {
             if(subCaptionText != null) marker.setSubCaptionText((String) subCaptionText);
 
             final Object subCaptionTextSize = json.get("subCaptionTextSize");
-            if (subCaptionTextSize != null) marker.setSubCaptionTextSize((int) subCaptionTextSize);
+            if (subCaptionTextSize != null) marker.setSubCaptionTextSize(Convert.toFloat(subCaptionTextSize));
 
             final Object subCaptionColor = json.get("subCaptionColor");
             if (subCaptionColor != null) marker.setSubCaptionColor(Convert.toColorInt(subCaptionColor));